requirements.txt
selenium 
webdriver-manager
fpdf
beautifulsoup4
requests
Pillow